Fix turbo-stream responses
There was a bug where the browser received empty turbo-stream responses. Not setting the layout to false for turbo-stream requests fixes this bug.

Surprisingly, this commit does not seem to re-introduce the bug that fef00cb fixes. I suspect it’s because I have since been passing in the correct `view_context`.

Moving on.